requirements.txt
asgiref==3.10.0
crispy-bootstrap4==2025.6
dj-database-url==3.0.1
Django==5.2.7
django-crispy-forms==2.4
django-filter==25.2
django-taggit==6.1.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
